plenty of improvements to position Latte panels
--used a video demonstrating multiple panels at the
bottom edge with padding between them
--max/minLength and offset are changed to float values
--max/minLength and offset can now be adjusted with
Ctrl+Wheel over their label in Appearance tab page. Clicking
the previous mentioned label will round the values
--introduce an autoSizeEnabled value in order to block
auto size mechanism when it becomes annoying. Such issue
appears more often to Latte panels
--block MenuZ from Latte icon heuristics

+++ app/view/view.cpp
@@ -920,12 +920,12 @@ QRect View::screenGeometry() const
     return QRect();
 }
 
-int View::offset() const
+float View::offset() const
 {
     return m_offset;
 }
 
-void View::setOffset(int offset)
+void View::setOffset(float offset)
 {
     if (m_offset == offset) {
         return;
